Complete flex exercise # 7
I enjoyed this challenge, one thing I struggled with is how to get all
the elements to occupy the full viewport height. Using margin-top: auto
on the footer wasn't cutting it, nor trying to use justify/align
flex-end.

It took using flex-grow: 1; on the .main-content to really take up all
that space. Will follow up on the soution to compare & contrast :)